Added ci, and mockup readme, unit tests, changelog

+++ input/rtlsdr/rtl_sdr_vita49_vrt_minimal_iq_sender.py
@@ -12,9 +12,9 @@ from rtlsdr import RtlSdr
 UDP_IP = "127.0.0.1"
 UDP_PORT = 5005
 BUF_SAMPLES = 2048
-SAMPLE_RATE = 2e6   # Hz
-CENTER_FREQ = 100e6     # Hz
-GAIN = 'auto'
+SAMPLE_RATE = 2e6  # Hz
+CENTER_FREQ = 100e6  # Hz
+GAIN = "auto"
 STREAM_ID = uuid.uuid4().bytes[:16]  # 16-byte unik ID
 
 # -----------------------------
@@ -22,29 +22,34 @@ STREAM_ID = uuid.uuid4().bytes[:16]  # 16-byte unik ID
 # -----------------------------
 sock = socket.socket(socket.AF_INET, socket.SOCK_DGRAM)
 
+
 # -----------------------------
 # VITA-49 IF Header funktioner
 # -----------------------------
 def now_vita_timestamp():
     t = time.time()
     sec = int(t)
-    frac = int((t - sec) * (1<<64))
+    frac = int((t - sec) * (1 << 64))
     return sec, frac
+
 
 def build_vita49_if_packet(iq_samples, pkt_count):
     """Skapar VITA-49 IF-paket med float32 IQ och metadata korrekt 32-byte header"""
     # Float32 I/Q, little-endian
-    iq_float32 = np.empty(len(iq_samples)*2, dtype='<f4')
-    iq_float32[0::2] = np.real(iq_samples).astype('<f4')
-    iq_float32[1::2] = np.imag(iq_samples).astype('<f4')
+    iq_float32 = np.empty(len(iq_samples) * 2, dtype="<f4")
+    iq_float32[0::2] = np.real(iq_samples).astype("<f4")
+    iq_float32[1::2] = np.imag(iq_samples).astype("<f4")
 
     pkt_type = 0x1
-    hdr_word0 = 0x40000000 | (pkt_type<<28) | (pkt_count & 0x0FFFFFFF)
+    hdr_word0 = 0x40000000 | (pkt_type << 28) | (pkt_count & 0x0FFFFFFF)
 
     # 32-byte header: word0, stream_id, pkt_count, sample_rate, center_freq
-    header = struct.pack('<I16sIff', hdr_word0, STREAM_ID, pkt_count, SAMPLE_RATE, CENTER_FREQ)
+    header = struct.pack(
+        "<I16sIff", hdr_word0, STREAM_ID, pkt_count, SAMPLE_RATE, CENTER_FREQ
+    )
 
     return header + iq_float32.tobytes()
+
 
 # -----------------------------
 # RTL-SDR setup
